LRAC-11336 Fix spotbugs

+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/java/com/liferay/osb/faro/web/internal/model/display/main/FaroSubscriptionDisplay.java
@@ -202,7 +202,7 @@ public class FaroSubscriptionDisplay {
 	}
 
 	private long _computeLimit(
-		int addOnLimit, long currentLimit, int quantity) {
+		long addOnLimit, long currentLimit, long quantity) {
 
 		if (currentLimit < 0) {
 			return currentLimit;
